requirements.txt
chromadb==1.1.1
transformers==4.57.0
python-dotenv==1.1.0
google-adk
flask
PyPDF2
unstructured[pdf]
nltk
scikit-learn
sentence-transformers
FlagEmbedding
litellm
a2a
a2a-sdk[all]
urllib3 < 2.4.0
langchain_text_splitters
langchain_openai
langchain_ollama
prometheus_client
fastmcp==2.12.5
streamlit
